Move TransferFileBytesRequestObserver
