+++ src/Catalyst.Cli/Config/shell.components.json
@@ -248,7 +248,7 @@
             ]
         },
         {
-            "type": "Catalyst.Common.FileTransfer.TransferFileBytesRequestObserver, Catalyst.Common",
+            "type": "Catalyst.Common.IO.Observables.TransferFileBytesRequestObserver, Catalyst.Common",
             "services": [
                 {
                     "type": "Catalyst.Common.Interfaces.IO.Observables.IRpcRequestObserver, Catalyst.Common"
