+++ src/Catalyst.Common.UnitTests/IO/Observables/TransferFileBytesRequestObserverTest.cs
@@ -26,12 +26,12 @@ using System.Linq;
 using System.Threading.Tasks;
 using Catalyst.Common.Config;
 using Catalyst.Common.Extensions;
-using Catalyst.Common.FileTransfer;
 using Catalyst.Common.Interfaces.FileTransfer;
 using Catalyst.Common.Interfaces.IO.Messaging;
 using Catalyst.Common.Interfaces.IO.Messaging.Dto;
 using Catalyst.Common.IO.Messaging;
 using Catalyst.Common.IO.Messaging.Dto;
+using Catalyst.Common.IO.Observables;
 using Catalyst.Protocol.Rpc.Node;
 using Catalyst.TestUtils;
 using DotNetty.Transport.Channels;
@@ -41,7 +41,7 @@ using NSubstitute;
 using Serilog;
 using Xunit;
 
-namespace Catalyst.Common.UnitTests.FileTransfer
+namespace Catalyst.Common.UnitTests.IO.Observables
 {
     public sealed class TransferFileBytesRequestObserverTest
     {
