+++ src/Catalyst.Node.Core.IntegrationTests/Modules/Dfs/NodeFileTransferIntegrationTest.cs
@@ -34,6 +34,7 @@ using Catalyst.Common.Interfaces.Modules.Dfs;
 using Catalyst.Common.Interfaces.Rpc;
 using Catalyst.Common.IO.Messaging;
 using Catalyst.Common.IO.Messaging.Dto;
+using Catalyst.Common.IO.Observables;
 using Catalyst.Common.P2P;
 using Catalyst.Node.Core.Modules.Dfs;
 using Catalyst.Node.Core.P2P;
