+++ src/Catalyst.Node.Core/Config/messageHandlers.json
@@ -67,7 +67,7 @@
             ]
         },
         {
-            "type": "Catalyst.Common.FileTransfer.TransferFileBytesRequestObserver, Catalyst.Common",
+            "type": "Catalyst.Common.IO.Observables.TransferFileBytesRequestObserver, Catalyst.Common",
             "services": [
                 {
                     "type": "Catalyst.Common.Interfaces.IO.Observables.IRpcRequestObserver, Catalyst.Common"
